+++ Sessions.MVP/Presenters/MarkersPresenter.cs
@@ -141,7 +141,7 @@ namespace Sessions.MVP.Presenters
                     Marker marker = new Marker();
                     marker.Name = markerName;
                     marker.PositionBytes = _playerService.GetPosition().PositionBytes;
-                    marker.PositionSamples = (uint)ConvertAudio.ToPCM(marker.PositionBytes, _playerService.CurrentPlaylistItem.AudioFile.BitsPerSample, 2);
+                    marker.PositionSamples = ConvertAudio.ToPCM(marker.PositionBytes, _playerService.CurrentPlaylistItem.AudioFile.BitsPerSample, 2);
                     long ms = ConvertAudio.ToMS(marker.PositionSamples, _playerService.CurrentPlaylistItem.AudioFile.SampleRate);
                     marker.Position = ConvertAudio.ToTimeString(ms);
                     marker.AudioFileId = _playerService.CurrentPlaylistItem.AudioFile.Id;
@@ -287,7 +287,7 @@ namespace Sessions.MVP.Presenters
                     // Update marker and update view
                     marker.Position = positionString;
                     marker.PositionBytes = positionBytes;
-                    marker.PositionSamples = (uint)positionSamples;
+                    marker.PositionSamples = positionSamples;
                     marker.PositionPercentage = positionPercentage;
                     _messageHub.PublishAsync<MarkerPositionUpdatedMessage>(new MarkerPositionUpdatedMessage(this, marker));
